refactor gcs bucket
